Fix Atomic generator bug where pre-commit files were ignored if the post-commit file exists
If the step chooses to generate the pre-commit file we should always use
that. If it _doesn't_ generate the file then we want to use the post-commit
file if it exists and throw an error otherwise.

Also added some debug logging to indicate that the pre-commit file is being
renamed to the post-commit name.